Additional distributions

AugmentedGPLikelihoods.SpecialDistributions.PolyaGamma — Type

PolyaGamma(b::Real, c::Real) <: ContinuousUnivariateDistribution

Arguments

b::Real

c::Real exponential tilting

Create a PolyaGamma sampler with parameters b and c. Note that sampling will differ if b is a Int or a Real.

AugmentedGPLikelihoods.SpecialDistributions.PolyaGammaPoisson — Type

PolyaGammaPoisson(y::Real, c::Real, λ::Real)

A bivariate distribution, used as hierachical prior as:

\[ p(\omega, n) = \operatorname{PG}(\omega|y + n, c)\operatorname{Po}(n|\lambda).\]

Random samples as well as statistics from the distribution will returned as NamedTuple : (;ω, n).

This structured distributions is needed for example for the PoissonLikelihood.

Additional likelihoods

Missing docstring.

Missing docstring for LaplaceLikelihood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NegBinomialLikelihood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StudentTLikelihood. Check Documenter's build log for details.

NamedTuple/TupleVector distribution interface

AugmentedGPLikelihoods.SpecialDistributions.AbstractNTDist — Type

AbstractNTDist is an abstract type for a wrapper type around measure(s) and distributions(s). The main idea is that instead of rand, mean and other statistical tools wrapped objects return NamedTuples or TupleVector when having a collection of them.

The following API has to be implemented: Given π::AbstractNTDist and Π::AbstractVector{<:AbstractNTDist}

Necessary

ntrand(rng, π) -> NamedTuple

ntmean(π) -> NamedTuple

MeasureBase.logdensity(π, x::NamedTuple) -> Real

Distributions.kldivergence(π₀, π₁) -> Real

Optional

tvrand(rng, Π) -> TupleVector

tvmean(Π) -> TupleVector

AugmentedGPLikelihoods.SpecialDistributions.NTDist — Type

NTDist(d) -> NTDist{typeof(d),:ω}
NTDist{S}(d) -> NTDist{typeof(d),S}

Wrapper around a single distribution to be compatible with the ntrand, ntmean interface. One can pass the wanted symbol via S while the default will be :ω.

AugmentedGPLikelihoods.SpecialDistributions.ntrand — Function

ntrand([rng::AbstractRNG,] d) -> NamedTuple

Return a sample as a NamedTuple.

AugmentedGPLikelihoods.SpecialDistributions.ntmean — Function

ntmean(d::Distribution) -> NamedTuple

Return the mean as a NamedTuple.

AugmentedGPLikelihoods.SpecialDistributions.tvrand — Function

tvrand([rng::AbstractRNG,] d::ProductMeasure) -> TupleVector
tvrand([rng::AbstractRNG,] d::AbstractVector{<:AbstractNTDist}) -> TupleVector

Return a collection of samples as a TupleVector

AugmentedGPLikelihoods.SpecialDistributions.tvmean — Function

tvmean(d::AbstractVector{<:AbstractNTDist}) -> TupleVector
tvmean(d::ProductMeasure)

Return a collection of mean as a TupleVector.